refactor: move some functions to unpopulateScript
Extract some functions into unpopulateScript to be used inside other
files. This reduces code duplications and allows the functions to be used
by other unpopulate scripts such as unpopulateClient script.
